Add accuracy_score metric and some trivial classifiers
This commit introduces the simple classifiers `TrivialClassifier` and `MajorityClassifier`. The `TrivialClassifier` simply predicts the class that it is told to in ::new(). This does not require fitting because it does not learn from the data at all. The `MajorityClassifier` is slightly more advanced and calculates the class with the highest frequency.

These both use the `Classifier` trait which represents a typical classification setup with a matrix where each row is a different sample with features in columns. This is likely to need revision over time, but for now is an adequate start.

The `accuracy_score` metric calculates the proportion of predictions which are correct.