Add GradientDescent solver, move OpenBLAS to dev dependency.
This commit adds the `GradientDescent` solver intended to perform the (non-stochastic) gradient descent algorithm for a function f. Usage documentation is given in the code.

The OpenBLAS dependency has been moved to the dev-dependencies as it is only required to run the tests. The library itself should not be linked directly to any linear algebra package to allow the consumer of the library to choose.